refactor: Execution클래스 삭제
inputView에 있어야할 결과출력이 Execution클래스에 있어 Execution클래스를 삭제했다.
