+++ src/main/java/kr/codesquad/controller/InputOutputController.java
@@ -1,22 +1,22 @@
 package kr.codesquad.controller;
 
-import kr.codesquad.domain.*;
-import kr.codesquad.view.ExecutionResult;
+import kr.codesquad.domain.Ladder;
+import kr.codesquad.domain.LadderHeight;
+import kr.codesquad.domain.Players;
+import kr.codesquad.domain.Results;
 import kr.codesquad.view.InputView;
 import kr.codesquad.view.OutputView;
 
-import java.io.IOException;
 import java.util.HashMap;
-import java.util.Optional;
 
 public class InputOutputController {
 
-    private OutputView outputHandler;
-    private InputView inputHandler;
+    private OutputView outputView;
+    private InputView inputView;
 
     public InputOutputController() {
-        this.inputHandler = new InputView();
-        this.outputHandler = new OutputView();
+        this.inputView = new InputView();
+        this.outputView = new OutputView();
     }
 
     void setupGame(Players players, LadderHeight ladderHeight, Results result) {
@@ -26,28 +26,29 @@ public class InputOutputController {
     }
 
     private void getPlayers(Players players) {
-        inputHandler.playerNamePrompt();
-        while (!players.getValidPlayerFromUser(inputHandler.getInput())) {
+        inputView.playerNamePrompt();
+        while (!players.getValidPlayerFromUser(inputView.getInput())) {
         }
     }
 
     private void getHeight(LadderHeight ladderHeight) {
-        inputHandler.LadderHeightPrompt();
-        while (!ladderHeight.getValidHeightFromUser(Integer.parseInt(inputHandler.getInput()))) ;
+        inputView.LadderHeightPrompt();
+        while (!ladderHeight.getValidHeightFromUser(Integer.parseInt(inputView.getInput()))) ;
     }
 
     private void getResults(Results result, int playerNum) {
-        inputHandler.resultPrompt();
-        while (!result.getResultsFromUser(inputHandler.getInput(), playerNum)) ;
+        inputView.resultPrompt();
+        while (!result.getResultsFromUser(inputView.getInput(), playerNum)) ;
     }
 
-    void getExecutionResult(HashMap<String, String> resultMap) throws IOException {
+    void getExecutionResult(HashMap<String, String> resultMap){
         while (true) {
-            new ExecutionResult(inputHandler, resultMap);
+            inputView.selectPlayerPrompt();
+            outputView.getExecutionResult(inputView.getInput(),resultMap);
         }
     }
 
     void printLadder(Ladder ladder, Players player, Results result) {
-        outputHandler.printLadder(ladder, player, result);
+        outputView.printLadder(ladder, player, result);
     }
 }
