+++ src/main/java/kr/codesquad/domain/Players.java
@@ -59,6 +59,8 @@ public class Players {
     }
 
 
+
+
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
